# web/openapi/mengnex.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/photos": {
      "get": {
        "tags": [
          "photos"
        ],
        "operationId": "list_photos",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of photos",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of photos to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "before_id",
            "in": "query",
            "description": "Return photos after this cursor id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "library_id",
            "in": "query",
            "description": "Limit results to one media library",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List scanned photos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhotoAssetResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/photos/folders/{library_id}": {
      "get": {
        "tags": [
          "photos"
        ],
        "operationId": "list_folder_contents",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "description": "Media library ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "Relative folder path",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum direct photos",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of direct photos to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Folder contents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoFolderContentsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/photos/{photo_id}": {
      "delete": {
        "tags": [
          "photos"
        ],
        "operationId": "delete_photo",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "description": "Photo asset ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Move photo asset to recycle bin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePhotoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "operationId": "get_preferences",
        "responses": {
          "200": {
            "description": "Application preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "preferences"
        ],
        "operationId": "update_preferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated application preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeletePhotoResponse": {
        "type": "object",
        "required": [
          "id",
          "file_id",
          "item_id",
          "source_path"
        ],
        "properties": {
          "file_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "item_id": {
            "type": "string"
          },
          "source_path": {
            "type": "string"
          }
        }
      },
...
      "PhotoAssetResponse": {
        "type": "object",
        "required": [
          "id",
          "item_id",
          "file_id",
          "library_id",
          "title",
          "file_name",
          "src",
          "original_src",
          "source_path",
          "file_size",
          "batch_time"
        ],
        "properties": {
          "batch_time": {
            "type": "string",
            "format": "date-time"
          },
          "file_id": {
            "type": "string"
          },
          "file_name": {
            "type": "string"
          },
          "file_size": {
            "type": "integer",
            "format": "int64"
          },
          "height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "id": {
            "type": "string"
          },
          "item_id": {
            "type": "string"
          },
          "library_id": {
            "type": "string"
          },
          "mime_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "original_src": {
            "type": "string"
          },
          "preview_src": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_path": {
            "type": "string"
          },
          "src": {
            "type": "string"
          },
          "taken_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "title": {
            "type": "string"
          },
          "width": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
      "PhotoFolderContentsResponse": {
        "type": "object",
        "required": [
          "library_id",
          "path",
          "folders",
          "photos",
          "total_photos"
        ],
        "properties": {
          "folders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhotoFolderResponse"
            }
          },
          "library_id": {
            "type": "string"
          },
          "next_offset": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "path": {
            "type": "string"
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhotoAssetResponse"
            }
          },
          "total_photos": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PhotoFolderResponse": {
        "type": "object",
        "required": [
          "name",
          "path",
          "photo_count"
        ],
        "properties": {
          "cover": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PhotoAssetResponse"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "photo_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PreferencesResponse": {
        "type": "object",
        "required": [
          "preview_max_dimension",
          "preview_quality",
          "media_cache_max_bytes",
          "video_probe_enabled",
          "video_probe_command",
          "video_probe_timeout_seconds",
          "video_ffmpeg_command",
          "video_cover_time_percent",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "media_cache_directory": {
            "type": [
              "string",
              "null"
            ]
          },
          "media_cache_max_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "preview_max_dimension": {
            "type": "integer",
            "format": "int32"
          },
          "preview_quality": {
            "type": "integer",
            "format": "int32"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "video_cover_time_percent": {
            "type": "integer",
            "format": "int32"
          },
          "video_ffmpeg_command": {
            "type": "string"
          },
          "video_probe_command": {
            "type": "string"
          },
          "video_probe_enabled": {
            "type": "boolean"
          },
          "video_probe_timeout_seconds": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UpdatePreferencesRequest": {
        "type": "object",
        "properties": {
          "media_cache_directory": {
            "type": [
              "string",
              "null"
            ]
          },
          "media_cache_max_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "preview_max_dimension": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "preview_quality": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "video_cover_time_percent": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "video_ffmpeg_command": {
            "type": [
              "string",
              "null"
            ]
          },
          "video_probe_command": {
            "type": [
              "string",
              "null"
            ]
          },
          "video_probe_enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "video_probe_timeout_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}